refactor(#342): extract RPC codec to core.rpc_codec — fix proxy→server cross-layer import
proxy/transport.py was importing decode_rpc_message from server.protocol,
violating KERNEL-ARCHITECTURE.md layering rules. Extracted all codec
functions (RPCEncoder, encode/decode_rpc_message, rpc_decode_hook, etc.)
into a new shared module nexus.core.rpc_codec and updated all 9 consumers.
