+++ src/nexus/core/rpc_transport.py
@@ -122,14 +122,9 @@ class NexusRPCTransport:
             TransportError: Connection or timeout error
             RPCError: Server returned an error response
         """
-        # Lazy import to avoid circular dependency
-        from nexus.server.protocol import (
-            RPCErrorCode,
-            RPCRequest,
-            RPCResponse,
-            decode_rpc_message,
-            encode_rpc_message,
-        )
+        # Import codec from shared core module; protocol types from server
+        from nexus.core.rpc_codec import decode_rpc_message, encode_rpc_message
+        from nexus.server.protocol import RPCErrorCode, RPCRequest, RPCResponse
 
         request = RPCRequest(
             method=method,
