+++ src/nexus/remote/async_client.py
@@ -57,14 +57,10 @@ from nexus.core.exceptions import (
     NexusFileNotFoundError,
 )
 from nexus.core.filesystem import NexusFilesystem
+from nexus.core.rpc_codec import decode_rpc_message, encode_rpc_message
 from nexus.remote.base_client import BaseRemoteNexusFS
 from nexus.remote.rpc_proxy import RPCProxyBase
-from nexus.server.protocol import (
-    RPCRequest,
-    RPCResponse,
-    decode_rpc_message,
-    encode_rpc_message,
-)
+from nexus.server.protocol import RPCRequest, RPCResponse
 
 from .client import (
     _DOMAIN_METHOD_MAP,
