+++ src/nexus/server/rpc_server.py
@@ -36,6 +36,7 @@ from nexus.core.exceptions import (
 )
 from nexus.core.filters import is_os_metadata_file
 from nexus.core.nexus_fs import NexusFS
+from nexus.core.rpc_codec import decode_rpc_message, encode_rpc_message
 from nexus.core.virtual_views import (
     add_virtual_views_to_listing,
     get_parsed_content,
@@ -50,8 +51,6 @@ from nexus.server.protocol import (
     RPCErrorCode,
     RPCRequest,
     RPCResponse,
-    decode_rpc_message,
-    encode_rpc_message,
     parse_method_params,
 )
